docs: 更新全部文档至 v1.20
- README.md: 添加 v1.20 更新日志
- campus_blog.md: 版本号 v1.19 → v1.20，文档版本更新
- 前端开发文档.md: 后端版本 v1.19 → v1.20，文档版本更新
- CAMPUS_BLOG_ENHANCEMENT_PLAN.md: 版本号 v1.19 → v1.20

+++ src/main/java/com/example/edu_project/service/impl/TrendingServiceImpl.java
@@ -165,7 +165,7 @@ public class TrendingServiceImpl extends ServiceImpl<BlogTrendingMapper, BlogTre
 
         BlogTrending trending = new BlogTrending();
         trending.setPostId(postId);
-        trending.setScore(score);
+        trending.setScore((double)score);
         trending.setViewCount(post.getViewCount());
         trending.setLikeCount(post.getLikeCount());
         trending.setCommentCount(post.getCommentCount());
